Fixed bug preventing options menu from appearing
I found that the options menu object in the player prefab had its scale value inexplicably set to zero, preventing it from appearing, a bug that seems to happen with infuriating frequency due to strange quirks of the prefab editor. I quickly fixed this so it should work now.

I also turned the player spaceship model and cosmetics into a separate prefab, so it can be reused for non-gameplay purposes.