Turned player spaceship HUD into its own prefab
This will make it easier to modify.
As per instructions I have been given, I plan on adding some minigames to the project, that will require additional infrastructure such as more heads-up-display information. So I will need to add a lot more elements to the spaceship HUD, that will enable/disable as needed.