New 'Minigame' class
This is a base class for various types of minigames the player can participate in. The class features base functions and variables such as incrementing progression, win and lose states, and a function for triggering the start of the game. UnityEvents are also included to make custom effects easier. Abstract functions are also included to update the player's HUD based on the specific things happening in the minigame.