TimeValue struct for setting time
This struct acts as an easy way to set and view time values.
Generally when keeping track of time, code will have a single float representing seconds. It's very good when being processed by the game, but bad for developers who want to set time values, or players who want to see the current time value. This struct stores hours, minutes and seconds, and has easy properties for obtaining the values as a single seconds value for use by the game, and an easy string for returning the value neatly formatted in hours, minutes and seconds.