Reworked VariableHeadsUpDisplay into MinigameHeadsUpDisplay
I renamed the class to make more sense in the context of what it does.
I also added extra code for counter data (e.g. if the player must perform a set number of actions within the time limit), and added a property to automatically set up the HUD if a new minigame is activated.